requirements-dev.txt
-r requirements.txt
pytest>=7.4
pytest-cov>=4.1
mypy>=1.8
ruff>=0.6
httpx>=0.27
boto3-stubs[s3]>=1.35.0
